Rename build-watch-and-serve.js to tasks.js
